Removes the wrapper class DummyErrorBody
