+++ apis/api-mal/src/main/java/org/ccsds/moims/mo/mal/MOErrorException.java
@@ -23,11 +23,14 @@ package org.ccsds.moims.mo.mal;
 import java.io.Serializable;
 import org.ccsds.moims.mo.mal.structures.Attribute;
 import org.ccsds.moims.mo.mal.structures.UInteger;
+import org.ccsds.moims.mo.mal.transport.MALEncodedBody;
+import org.ccsds.moims.mo.mal.transport.MALEncodedElement;
+import org.ccsds.moims.mo.mal.transport.MALErrorBody;
 
 /**
  * Represents a MAL error.
  */
-public class MOErrorException extends Exception implements Serializable {
+public class MOErrorException extends Exception implements Serializable, MALErrorBody {
 
     private final UInteger errorNumber;
     private final Object extraInformation;
@@ -70,6 +73,31 @@ public class MOErrorException extends Exception implements Serializable {
      */
     public Object getExtraInformation() {
         return extraInformation;
+    }
+
+    @Override
+    public MOErrorException getError() throws MALException {
+        return this;
+    }
+
+    @Override
+    public int getElementCount() {
+        return 1;
+    }
+
+    @Override
+    public Object getBodyElement(int index, Object element) throws MALException {
+        return this;
+    }
+
+    @Override
+    public MALEncodedElement getEncodedBodyElement(int index) throws MALException {
+        return null;
+    }
+
+    @Override
+    public MALEncodedBody getEncodedBody() throws MALException {
+        return null;
     }
 
     @Override
